requirements.txt
coremltools==6.3.0
matplotlib==3.5.1
mss==6.1.0
numpy~=1.24.4
opencv_python==4.7.0.68
pandas==1.4.1
Pillow==9.5.0
psutil==5.9.1
PyQt5==5.15.9
PyQt5_sip==12.11.0
PyYAML==6.0
QDarkStyle==3.1
Requests==2.30.0
scipy==1.8.0
seaborn==0.11.2
thop==0.1.1.post2209072238
tqdm==4.64.1

opencv-python~=4.7.0.68
torch~=2.0.0+cu118
torchvision~=0.15.1+cu118
ui~=0.1.4
reportlab~=4.0.4